Lots of changes.
Renamed repo to powerpc-wrs-vxworks-buildscripts
Merged in vxworks-build-helper
Made cmake file to build/install everything
Made the toolchain easier to configure, and turned it into a template
that generates both a toolchain file and a platform file for "FRC" -
didn't want to use vxWorks as it seems like this configuration is too
specific to say it is a generic vxworks toolchain configuration.
Removed FindWPILib as it should be with WPILib, not the project
Added debian stuff.
Added a deploy macro for FRC projects